+++ test/test_hsp.py
@@ -64,7 +64,7 @@ class TestHSP(TestCase):
                   subjectStart=5, subjectEnd=6,
                   readMatchedSequence='aaa', subjectMatchedSequence='ccc',
                   readFrame=7, subjectFrame=8, identicalCount=9,
-                  positiveCount=10)
+                  positiveCount=10, percentIdentical=99.3)
 
         self.assertEqual(
             {
@@ -81,6 +81,7 @@ class TestHSP(TestCase):
                 'positiveCount': 10,
                 'readMatchedSequence': 'aaa',
                 'subjectMatchedSequence': 'ccc',
+                'percentIdentical': 99.3,
             },
             hsp.toDict())
 
@@ -146,7 +147,7 @@ class TestLSP(TestCase):
                   subjectStart=5, subjectEnd=6,
                   readMatchedSequence='aaa', subjectMatchedSequence='ccc',
                   readFrame=7, subjectFrame=8, identicalCount=9,
-                  positiveCount=10)
+                  positiveCount=10, percentIdentical=99.3)
 
         self.assertEqual(
             {
@@ -163,5 +164,6 @@ class TestLSP(TestCase):
                 'positiveCount': 10,
                 'readMatchedSequence': 'aaa',
                 'subjectMatchedSequence': 'ccc',
+                'percentIdentical': 99.3,
             },
             lsp.toDict())