+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.1.22 Dec 3, 2019
+
+Added `--percentageIdenticalCutoff` argument to `alignment-panel-civ.py`
+and `noninteractive-alignment-panel.py` and all that implies, down to
+reading DIAMOND output with the `pident` value, storing it, restoring it,
+and filtering on it in `dark.alignments.ReadsAlignmentsFilter`, plus tests.
+
 ## 3.1.21 Nov 30, 2019
 
 Minor changes to HTML output.